feat: Update dependencies to allow PHP 7.4 and 8.0 & fix compatibilty error with sylius
