+++ .github/workflows/tests.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches: [ master ]
   pull_request:
-    branches: [ master ]
   schedule:
     - cron:  '0 0 * * *'
 
@@ -17,31 +16,37 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        php: [7.2, 7.3]
-        prefer: [lowest, stable]
+        php: ['7.4', '8.0']
+        prefer: ['lowest', 'stable']
 
     steps:
       - uses: actions/checkout@v2
 
       - name: Switch PHP
-        run: sudo update-alternatives --set php /usr/bin/php${{ matrix.php }}
-
-      - name: Show PHP version
-        run: php -v
+        run: |
+          sudo update-alternatives --set php /usr/bin/php${{ matrix.php }}
+          echo ${{ matrix.php }} > .php-version
+          php -v
 
       - uses: actions/cache@v1
         with:
           path: /home/runner/.composer/cache
-          key: composer-php.${{ matrix.php }}-prefer.${{ matrix.prefer }}-${{ github.sha }}
-          restore-keys: composer-php.${{ matrix.php }}-prefer.${{ matrix.prefer }}-
+          key: composer2-php:${{ matrix.php }}-${{ github.sha }}
+          restore-keys: composer2-php:${{ matrix.php }}-
 
-      - name: Update composer
+      - run: mkdir -p /home/runner/{.composer/cache,.config/composer}
+        if: steps.cache-composer.outputs.cache-hit != 'true'
+
+      - name: Composer v2
+        run: sudo composer self-update --2
+
+      - name: Install symfony CLI
         run: |
-          mkdir -p /home/runner/.composer/
-          sudo composer self-update --1
+          curl https://get.symfony.com/cli/installer | bash
+          echo ${HOME}/.symfony/bin >> $GITHUB_PATH
 
       - name: Validate composer.json and composer.lock
-        run: composer validate --strict
+        run: composer validate --strict --no-check-all
 
       - name: Composer Github Auth
         run: composer config -g github-oauth.github.com ${{ github.token }}
@@ -57,34 +62,31 @@ jobs:
 
     runs-on: ubuntu-latest
 
-    env:
-      SYMFONY_ENDPOINT: http://127.0.0.1/
-
     strategy:
       fail-fast: false
       matrix:
-        php: [7.2, 7.3]
-        sylius: [1.3, 1.4, 1.5, 1.6, 1.7]
-        mariadb_version: [10.3, 10.4]
+        php: ['7.4', '8.0']
+        sylius: ['1.8', '1.9', '1.10']
+        mariadb_version: ['10.4']
+        flex: ["standalone", "official"]
         exclude:
-          - php: 7.2
-            sylius: 1.7
+          - php: '8.0'
+            sylius: '1.8'
+          - php: '8.0'
+            sylius: '1.9'
 
     steps:
-      - name: Remove HHVM
-        run: sudo apt-get remove --auto-remove -y hhvm
-        if: matrix.sylius == 1.3 || matrix.sylius == 1.4
+      - name: Define SYMFONY_ENDPOINT
+        if: matrix.flex == 'standalone'
+        run: |
+          echo "SYMFONY_ENDPOINT=http://127.0.0.1/" >> $GITHUB_ENV
 
       - name: Switch PHP
-        run: sudo update-alternatives --set php /usr/bin/php${{ matrix.php }}
-
-      - name: Show PHP version
-        run: php -v
-
-      - name: Setup timezone
         run: |
-          echo "date.timezone=UTC" >> /tmp/timezone.ini; \
+          sudo update-alternatives --set php /usr/bin/php${{ matrix.php }}
+          echo "date.timezone=UTC" >> /tmp/timezone.ini
           sudo mv /tmp/timezone.ini /etc/php/${{ matrix.php }}/cli/conf.d/timezone.ini
+          echo ${{ matrix.php }} > .php-version
 
       - uses: actions/checkout@v2
         with:
@@ -92,6 +94,7 @@ jobs:
 
       # Run the server at the start so it can download the recipes!
       - name: Run standalone symfony flex server
+        if: matrix.flex == 'standalone'
         run: |
           echo ${{ github.token }} | docker login docker.pkg.github.com -u ${{ github.actor }} --password-stdin
           docker run --rm --name flex -d -v $PWD/plugin/recipe:/var/www/flex/var/repo/private/monsieurbiz/sylius-alert-message-plugin -p 80:80 docker.pkg.github.com/monsieurbiz/docker/symfony-flex-server:latest contrib official
@@ -100,13 +103,19 @@ jobs:
       - uses: actions/cache@v1
         with:
           path: /home/runner/.composer/cache
-          key: composer-php.${{ matrix.php }}-sylius.${{ matrix.sylius }}.0-${{ github.sha }}
-          restore-keys: composer-php.${{ matrix.php }}-sylius.${{ matrix.sylius }}.0-
+          key: composer2-php:${{ matrix.php }}-${{ github.sha }}
+          restore-keys: composer2-php:${{ matrix.php }}-
 
-      - name: Update composer
+      - run: mkdir -p /home/runner/{.composer/cache,.config/composer}
+        if: steps.cache-composer.outputs.cache-hit != 'true'
+
+      - name: Composer v2
+        run: sudo composer self-update --2
+
+      - name: Install symfony CLI
         run: |
-          mkdir -p /home/runner/.composer/
-          sudo composer self-update --1
+          curl https://get.symfony.com/cli/installer | bash
+          echo ${HOME}/.symfony/bin >> $GITHUB_PATH
 
       - name: Composer Github Auth
         run: composer config -g github-oauth.github.com ${{ github.token }}
@@ -121,18 +130,13 @@ jobs:
           php bin/console doctrine:database:create --if-not-exists
           php bin/console doctrine:migr:migr -n
 
-      - name: Add path repository
+      - name: Setup some requirements
         working-directory: ./sylius
         run: |
           composer config repositories.plugin '{"type": "path", "url": "../plugin/"}'
-
-      - name: Allow contrib recipes
-        working-directory: ./sylius
-        run: composer config extra.symfony.allow-contrib true
-
-      - name: Allow unsecure localhost flex server
-        working-directory: ./sylius
-        run: composer config secure-http false
+          composer config extra.symfony.allow-contrib true
+          composer config secure-http false
+          composer config --unset platform.php
 
       - name: Install plugin
         working-directory: ./sylius
@@ -140,6 +144,7 @@ jobs:
           composer require monsieurbiz/sylius-alert-message-plugin="*@dev"
 
       - name: Show flex server logs
+        if: matrix.flex == 'standalone'
         run: docker logs --tail 100 flex
 
       - name: Update database schema
