+++ .php-version.dist
@@ -1,0 +1,1 @@
+7.4
