+++ composer.json
@@ -1,38 +1,38 @@
 {
     "name": "monsieurbiz/sylius-alert-message-plugin",
     "type": "sylius-plugin",
-    "keywords": ["sylius", "sylius-plugin"],
+    "keywords": ["sylius", "sylius-plugin", "alert", "monsieurbiz"],
     "description": "Adds a simple message on top of your shop.",
     "license": "MIT",
     "require": {
-        "php": "~7.2, <7.4",
-        "sylius/sylius": "~1.3",
-        "gedmo/doctrine-extensions": "^2.4.12"
+        "php": "~7.4 || ^8.0",
+        "sylius/sylius": "^1.8",
+        "gedmo/doctrine-extensions": "^2.4.12 || ^3.0"
     },
     "require-dev": {
-        "behat/behat": "^3.4",
-        "behat/mink": "^1.7@dev",
-        "behat/mink-browserkit-driver": "^1.3",
-        "behat/mink-extension": "^2.2",
-        "behat/mink-selenium2-driver": "^1.3",
+        "behat/behat": "^3.6.1",
+        "behat/mink-selenium2-driver": "^1.4",
+        "friends-of-behat/mink": "^1.8",
+        "friends-of-behat/mink-browserkit-driver": "^1.4",
+        "friends-of-behat/mink-extension": "^2.4",
         "friends-of-behat/page-object-extension": "^0.3",
         "friends-of-behat/suite-settings-extension": "^1.0",
         "friends-of-behat/symfony-extension": "^2.0",
         "friends-of-behat/variadic-extension": "^1.1",
-        "lakion/mink-debug-extension": "^1.2.3",
-        "phpspec/phpspec": "^5.0",
-        "phpstan/phpstan-doctrine": "^0.11",
-        "phpstan/phpstan-shim": "^0.11",
-        "phpstan/phpstan-webmozart-assert": "^0.11",
-        "phpunit/phpunit": "^8.0",
-        "sensiolabs/security-checker": "^5.0",
-        "sylius-labs/coding-standard": "^3.0",
-        "symfony/browser-kit": "^3.4|^4.3",
-        "symfony/debug-bundle": "^3.4|^4.3",
-        "symfony/dotenv": "^4.3",
-        "symfony/intl": "^3.4|^4.3",
-        "symfony/web-profiler-bundle": "^3.4|^4.3",
-        "symfony/web-server-bundle": "^3.4|^4.3"
+        "lchrusciel/api-test-case": "^3.0 || ^5.0",
+        "phpspec/phpspec": "^7.0",
+        "phpstan/phpstan-doctrine": "0.12.33",
+        "phpstan/phpstan": "^0.12",
+        "phpstan/phpstan-webmozart-assert": "^0.12.7",
+        "phpunit/phpunit": "^8.5",
+        "sensiolabs/security-checker": "^6.0",
+        "sylius-labs/coding-standard": "^4.0",
+        "symfony/browser-kit": "^4.4 || ^5.2",
+        "symfony/debug-bundle": "^4.4 || ^5.2",
+        "symfony/dotenv": "^4.4 || ^5.2",
+        "symfony/intl": "^4.4 || ^5.2",
+        "symfony/web-profiler-bundle": "^4.4 || ^5.2",
+        "symfony/web-server-bundle": "^4.4 || ^5.2"
     },
     "prefer-stable": true,
     "autoload": {
@@ -41,12 +41,15 @@
         }
     },
     "autoload-dev": {
+        "classmap": ["tests/Application/src/Kernel.php"],
         "psr-4": {
-            "Tests\\MonsieurBiz\\SyliusAlertMessagePlugin\\": "tests/"
-        },
-        "classmap": ["tests/Application/Kernel.php"]
+            "App\\": "tests/Application/src/"
+        }
     },
     "extra": {
+        "symfony": {
+            "require": "^4.4"
+        },
         "branch-alias": {
             "dev-master": "1.0-dev"
         }
