+++ src/DependencyInjection/MonsieurBizSyliusAlertMessageExtension.php
@@ -7,16 +7,17 @@ namespace MonsieurBiz\SyliusAlertMessagePlugin\DependencyInjection;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
 
-final class MonsieurBizSyliusAlertMessageExtension extends Extension
+final class MonsieurBizSyliusAlertMessageExtension extends Extension implements PrependExtensionInterface
 {
     /**
      * {@inheritdoc}
      */
     public function load(array $config, ContainerBuilder $container): void
     {
-        $config = $this->processConfiguration($this->getConfiguration([], $container), $config);
+        $this->processConfiguration($this->getConfiguration([], $container), $config);
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yaml');
     }
@@ -24,5 +25,18 @@ final class MonsieurBizSyliusAlertMessageExtension extends Extension
     public function getAlias()
     {
         return str_replace('monsieur_biz', 'monsieurbiz', parent::getAlias());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function prepend(ContainerBuilder $container): void
+    {
+        $doctrineConfig = $container->getExtensionConfig('doctrine_migrations');
+        $container->prependExtensionConfig('doctrine_migrations', [
+            'migrations_paths' => array_merge(array_pop($doctrineConfig)['migrations_paths'] ?? [], [
+                'MonsieurBiz\SyliusAlertMessagePlugin\Migrations' => '@MonsieurBizSyliusAlertMessagePlugin/Migrations',
+            ]),
+        ]);
     }
 }
