+++ src/Helper/MessageHelper.php
@@ -1,19 +1,17 @@
 <?php
 declare(strict_types=1);
 
-namespace MonsieurBiz\SyliusAlertMessagePlugin\Block;
+namespace MonsieurBiz\SyliusAlertMessagePlugin\Helper;
 
+use MonsieurBiz\SyliusAlertMessagePlugin\Entity\Message;
 use MonsieurBiz\SyliusAlertMessagePlugin\Repository\MessageRepositoryInterface;
-use Sonata\BlockBundle\Event\BlockEvent;
+use Sonata\BlockBundle\Model\Block;
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Customer\Context\CustomerContextInterface;
 use Sylius\Component\Locale\Context\LocaleContextInterface;
 
-final class BlockEventListener
+final class MessageHelper extends Block
 {
-    /** @var string */
-    private $template;
-
     /**
      * @var ChannelContextInterface
      */
@@ -35,46 +33,58 @@ final class BlockEventListener
     private $messageRepository;
 
     /**
-     * BlockEventListener constructor.
+     * MessageHelper constructor.
      *
-     * @param string $template
      * @param ChannelContextInterface $channelContext
      * @param LocaleContextInterface $localeContext
      * @param CustomerContextInterface $customerContext
      * @param MessageRepositoryInterface $messageRepository
      */
     public function __construct(
-        string $template,
         ChannelContextInterface $channelContext,
         LocaleContextInterface $localeContext,
         CustomerContextInterface $customerContext,
         MessageRepositoryInterface $messageRepository
     ) {
-        $this->template = $template;
+        parent::__construct();
         $this->channelContext = $channelContext;
         $this->localeContext = $localeContext;
         $this->customerContext = $customerContext;
         $this->messageRepository = $messageRepository;
     }
 
-    /**
-     * @param BlockEvent $event
-     */
-    public function onBlockEvent(BlockEvent $event): void
+    public function getMessages(): array
     {
-        $block = new MessageBlock(
-            $this->channelContext,
-            $this->localeContext,
-            $this->customerContext,
-            $this->messageRepository
+        return $this->messageRepository->getActiveMessagesForChannelAndLocale(
+            $this->channelContext->getChannel(),
+            $this->localeContext->getLocaleCode()
         );
-        $block->setId(uniqid('', true));
-        $block->setSettings(array_replace($event->getSettings(), [
-            'template' => $this->template,
-        ]));
-        $block->setType('sonata.block.service.template');
+    }
 
-        $event->addBlock($block);
+    /**
+     * @param Message $message
+     *
+     * @return bool
+     */
+    public function canDisplayMessage(Message $message): bool
+    {
+        return
+            !$message->isCustomersOnly()
+            || null !== $this->customerContext->getCustomer()
+        ;
+    }
+
+    /**
+     * @param Message $message
+     *
+     * @return string|null
+     */
+    public function formatMessageWithTemplate(Message $message): ?string
+    {
+        $template = $message->getTemplateHtml();
+        $template = preg_replace('`{{\s*title\s*}}`i', $message->getTitle(), $template);
+        $template = preg_replace('`{{\s*message\s*}}`i', $message->getMessage(), $template);
+        return $template;
     }
 
 }
