+++ src/Resources/config/config.yaml
@@ -1,3 +1,4 @@
 imports:
   - { resource: "sylius/resources.yaml" }
   - { resource: "sylius/grid.yaml" }
+  - { resource: "sylius/ui_events.yaml" }
