+++ src/Resources/config/routing/admin.yaml
@@ -2,7 +2,7 @@ monsieurbiz_sylius_alert_message_admin:
   resource: |
     alias: monsieurbiz_alert_message.message
     section: admin
-    templates: SyliusAdminBundle:Crud
+    templates: "@SyliusAdmin\\Crud"
     except: ['show']
     redirect: update
     grid: monsieurbiz_alert_message
