+++ src/Resources/config/sylius/grid.yaml
@@ -23,12 +23,12 @@ sylius_grid:
           type: twig
           label: monsieurbiz_alert_message.ui.form.enabled
           options:
-            template: SyliusUiBundle:Grid/Field:enabled.html.twig
+            template: "@SyliusUi/Grid/Field/enabled.html.twig"
         customersOnly:
           type: twig
           label: monsieurbiz_alert_message.ui.form.customers_only
           options:
-            template: MonsieurBizSyliusAlertMessagePlugin:Grid/Field:customers_only.html.twig
+            template: "@MonsieurBizSyliusAlertMessagePlugin/Grid/Field/customers_only.html.twig"
       actions:
         main:
           create:
