+++ src/Resources/config/sylius_labs_doctrine_migrations_extra.yaml
@@ -1,0 +1,3 @@
+sylius_labs_doctrine_migrations_extra:
+    migrations:
+        'MonsieurBiz\SyliusAlertMessagePlugin\Migrations': ~
